Major overhaul
Expanded edit canvas, and had to retool a lot work correctly with it. I had to rework nearly the entire render pipeline, the object placement code, and the spawn placement code. I also added visualization of scroll data. This may be the longest I've spent on any one feature for this project to data.